+++ db/migrate/20230808152500_add_transfer_supplemental_funding_agreement_task_attributes.rb
@@ -1,0 +1,7 @@
+class AddTransferSupplementalFundingAgreementTaskAttributes < ActiveRecord::Migration[7.0]
+  def change
+    add_column :transfer_tasks_data, :supplemental_funding_agreement_received, :boolean
+    add_column :transfer_tasks_data, :supplemental_funding_agreement_cleared, :boolean
+    add_column :transfer_tasks_data, :supplemental_funding_agreement_saved, :boolean
+  end
+end